test(`cqn4sql`): regroup calc-element specs by content concern
The placement-based split (`select-list` vs `outside-select-list`)
mixed several distinct concerns inside a single 889-line file. Reorganize
into content-focused specs, each grouping tests by the unfolder aspect
they exercise:

- basics.spec.js              - trivial alias/val/scalar xpr
- nesting.spec.js             - calc element references calc element
- functions.spec.js           - function calls, named params, native fns
- ternary.spec.js             - CASE / ternary unfolding
- joins.spec.js               - join generation (incl. nested describe
                                for join-node reuse)
- filters.spec.js             - infix filters in / on calc elements
- exists.spec.js              - EXISTS-rejection + calc-elem-with-exists
- variable-replacements.spec.js - $now / $user preservation
- wildcard.spec.js            - {*} / excluding interactions
- nested-projections.spec.js  - inline / expand / subquery
- placements.spec.js          - where/from/group by/having/order by

Test bodies are moved verbatim; no semantics changed.
